refactor(Referee): Referee 에서 게임 결과만을 담당하는 GameResult 클래스 분리하여 생성
