+++ src/main/java/baseball/BaseballGame.java
@@ -1,5 +1,9 @@
 package baseball;
 
+import baseball.domain.Computer;
+import baseball.domain.Referee;
+import baseball.domain.GameResult;
+import baseball.domain.User;
 import baseball.view.Message;
 
 import static baseball.constant.Const.GAME_END_STRIKE_COUNT;
@@ -24,16 +28,19 @@ public class BaseballGame {
     }
 
     private void init() {
-        referee.initCount();
         computer.generateAnswer();
     }
 
     private void start() {
+        int strikeCount;
+
         do {
             user.enterNumber();
-            referee.judge(computer.getAnswer(), user.getInputNumber());
-            referee.showResult();
-        } while (referee.getStrikeCount() != GAME_END_STRIKE_COUNT);
+            GameResult gameResult = referee.judge(computer.getAnswer(), user.getInputNumber());
+            strikeCount = gameResult.getStrikeCount();
+            gameResult.show();
+
+        } while (strikeCount != GAME_END_STRIKE_COUNT);
     }
 
     private void end() {
